requirements.txt
Django==1.5.2
django-auth-ldap==1.1.4
django-debug-toolbar==0.9.4
django-extensions==1.1.1
django-ldapdb==0.1.0
python-ldap==2.4.13
six==1.3.0
wsgiref==0.1.2
